+++ src/test/java/org/opensearch/security/identity/SecurityTokenManagerTest.java
@@ -12,8 +12,10 @@
 package org.opensearch.security.identity;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.Set;
 
+import com.google.common.io.BaseEncoding;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -36,16 +38,15 @@ import org.opensearch.security.user.User;
 import org.opensearch.security.user.UserService;
 import org.opensearch.threadpool.ThreadPool;
 
+import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
@@ -75,11 +76,14 @@ public class SecurityTokenManagerTest {
 
     @After
     public void after() {
-        verifyNoMoreInteractions(cs);
-        verifyNoMoreInteractions(threadPool);
         verifyNoMoreInteractions(userService);
     }
 
+    final static String signingKey =
+        "This is my super safe signing key that no one will ever be able to guess. It's would take billions of years and the world's most powerful quantum computer to crack";
+    final static String signingKeyB64Encoded = BaseEncoding.base64().encode(signingKey.getBytes(StandardCharsets.UTF_8));
+
+    @Test
     public void onConfigModelChanged_oboNotSupported() {
         final ConfigModel configModel = mock(ConfigModel.class);
 
@@ -92,7 +96,7 @@ public class SecurityTokenManagerTest {
     @Test
     public void onDynamicConfigModelChanged_JwtVendorEnabled() {
         final ConfigModel configModel = mock(ConfigModel.class);
-        final DynamicConfigModel mockConfigModel = createMockJwtVendorInTokenManager();
+        final DynamicConfigModel mockConfigModel = createMockJwtVendorInTokenManager(true);
 
         tokenManager.onConfigModelChanged(configModel);
 
@@ -114,8 +118,12 @@ public class SecurityTokenManagerTest {
     }
 
     /** Creates the jwt vendor and returns a mock for validation if needed */
-    private DynamicConfigModel createMockJwtVendorInTokenManager() {
-        final Settings settings = Settings.builder().put("enabled", true).build();
+    private DynamicConfigModel createMockJwtVendorInTokenManager(boolean includeEncryptionKey) {
+        final Settings settings = Settings.builder()
+            .put("enabled", true)
+            .put("signing_key", signingKeyB64Encoded)
+            .put("encryption_key", (includeEncryptionKey ? "1234567890" : null))
+            .build();
         final DynamicConfigModel dcm = mock(DynamicConfigModel.class);
         when(dcm.getDynamicOnBehalfOfSettings()).thenReturn(settings);
         doAnswer((invocation) -> jwtVendor).when(tokenManager).createJwtVendor(settings);
@@ -207,11 +215,9 @@ public class SecurityTokenManagerTest {
         tokenManager.onConfigModelChanged(configModel);
         when(configModel.mapSecurityRoles(any(), any())).thenReturn(Set.of());
 
-        createMockJwtVendorInTokenManager();
+        createMockJwtVendorInTokenManager(true);
 
-        when(jwtVendor.createJwt(any(), anyString(), anyString(), anyLong(), any(), any(), anyBoolean())).thenThrow(
-            new RuntimeException("foobar")
-        );
+        when(jwtVendor.createJwt(any(), any(), any(), any())).thenThrow(new RuntimeException("foobar"));
         final OpenSearchSecurityException exception = assertThrows(
             OpenSearchSecurityException.class,
             () -> tokenManager.issueOnBehalfOfToken(null, new OnBehalfOfClaims("elmo", 450L))
@@ -233,15 +239,102 @@ public class SecurityTokenManagerTest {
         tokenManager.onConfigModelChanged(configModel);
         when(configModel.mapSecurityRoles(any(), any())).thenReturn(Set.of());
 
-        createMockJwtVendorInTokenManager();
+        createMockJwtVendorInTokenManager(true);
 
         final ExpiringBearerAuthToken authToken = mock(ExpiringBearerAuthToken.class);
-        when(jwtVendor.createJwt(any(), anyString(), anyString(), anyLong(), any(), any(), anyBoolean())).thenReturn(authToken);
+        when(jwtVendor.createJwt(any(), any(), any(), any())).thenReturn(authToken);
         final AuthToken returnedToken = tokenManager.issueOnBehalfOfToken(null, new OnBehalfOfClaims("elmo", 450L));
 
         assertThat(returnedToken, equalTo(authToken));
 
         verify(cs).getClusterName();
         verify(threadPool).getThreadContext();
+    }
+
+    @Test
+    public void testCreateJwtWithNegativeExpiry() {
+        doAnswer(invocation -> true).when(tokenManager).issueOnBehalfOfTokenAllowed();
+        final ThreadContext threadContext = new ThreadContext(Settings.EMPTY);
+        threadContext.putTransient(ConfigConstants.OPENDISTRO_SECURITY_USER, new User("Jon"));
+        when(threadPool.getThreadContext()).thenReturn(threadContext);
+        final ConfigModel configModel = mock(ConfigModel.class);
+        tokenManager.onConfigModelChanged(configModel);
+        when(configModel.mapSecurityRoles(any(), any())).thenReturn(Set.of());
+
+        createMockJwtVendorInTokenManager(true);
+
+        final Throwable exception = assertThrows(RuntimeException.class, () -> {
+            try {
+                final AuthToken returnedToken = tokenManager.issueOnBehalfOfToken(null, new OnBehalfOfClaims("elmo", -300L));
+            } catch (final Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+        assertThat(exception.getMessage(), is("java.lang.IllegalArgumentException: The expiration time should be a positive integer"));
+    }
+
+    @Test
+    public void testCreateJwtWithExceededExpiry() throws Exception {
+        doAnswer(invocation -> new ClusterName("cluster17")).when(cs).getClusterName();
+        doAnswer(invocation -> true).when(tokenManager).issueOnBehalfOfTokenAllowed();
+        final ThreadContext threadContext = new ThreadContext(Settings.EMPTY);
+        threadContext.putTransient(ConfigConstants.OPENDISTRO_SECURITY_USER, new User("Jon"));
+        when(threadPool.getThreadContext()).thenReturn(threadContext);
+        final ConfigModel configModel = mock(ConfigModel.class);
+        tokenManager.onConfigModelChanged(configModel);
+        when(configModel.mapSecurityRoles(any(), any())).thenReturn(Set.of());
+
+        createMockJwtVendorInTokenManager(true);
+
+        tokenManager.issueOnBehalfOfToken(null, new OnBehalfOfClaims("elmo", 90000000L));
+        // Expiry is a hint, the max value is controlled by the JwtVendor and reduced as is seen fit.
+        ArgumentCaptor<Long> longCaptor = ArgumentCaptor.forClass(Long.class);
+        verify(jwtVendor).createJwt(any(), any(), any(), longCaptor.capture());
+
+        assertThat(600L, equalTo(longCaptor.getValue()));
+    }
+
+    @Test
+    public void testCreateJwtWithBadEncryptionKey() {
+        doAnswer(invocation -> true).when(tokenManager).issueOnBehalfOfTokenAllowed();
+        final ThreadContext threadContext = new ThreadContext(Settings.EMPTY);
+        threadContext.putTransient(ConfigConstants.OPENDISTRO_SECURITY_USER, new User("Jon"));
+        when(threadPool.getThreadContext()).thenReturn(threadContext);
+        final ConfigModel configModel = mock(ConfigModel.class);
+        tokenManager.onConfigModelChanged(configModel);
+        when(configModel.mapSecurityRoles(any(), any())).thenReturn(Set.of());
+
+        createMockJwtVendorInTokenManager(false);
+
+        final Throwable exception = assertThrows(RuntimeException.class, () -> {
+            try {
+                tokenManager.issueOnBehalfOfToken(null, new OnBehalfOfClaims("elmo", 90000000L));
+            } catch (final Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+        assertThat(exception.getMessage(), is("java.lang.IllegalArgumentException: encryption_key cannot be null"));
+    }
+
+    @Test
+    public void testCreateJwtWithBadRoles() {
+        doAnswer(invocation -> true).when(tokenManager).issueOnBehalfOfTokenAllowed();
+        final ThreadContext threadContext = new ThreadContext(Settings.EMPTY);
+        threadContext.putTransient(ConfigConstants.OPENDISTRO_SECURITY_USER, new User("Jon"));
+        when(threadPool.getThreadContext()).thenReturn(threadContext);
+        final ConfigModel configModel = mock(ConfigModel.class);
+        tokenManager.onConfigModelChanged(configModel);
+        when(configModel.mapSecurityRoles(any(), any())).thenReturn(null);
+
+        createMockJwtVendorInTokenManager(true);
+
+        final Throwable exception = assertThrows(RuntimeException.class, () -> {
+            try {
+                tokenManager.issueOnBehalfOfToken(null, new OnBehalfOfClaims("elmo", 90000000L));
+            } catch (final Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+        assertThat(exception.getMessage(), is("java.lang.IllegalArgumentException: Roles cannot be null"));
     }
 }